import sbt project even if we fail to extract the scala instance due to any reason #SCL-24321
It's different from how it works in BSP, where it simply ignores sbt projects with any errors inside. However, it seems like the behaviour is also non-optimal as lots of modules end up missing in the IDE.

Overall, this change does not fix all the issues with the Scala 3 project structure importing in IntelliJ using SBT mode. There are still lots of questions - the repo is far from the standard. But at least the project can be imported in SBT mode now, even if the structure is not perfect.